Refactor to return reporter interface in livestatereporter
